refactor/#35-멤버 패키지 별도 생성 - 멤버 패키지 별도 생성
