+++ src/main/java/com/assu/server/domain/admin/entity/Admin.java
@@ -1,6 +1,6 @@
 package com.assu.server.domain.admin.entity;
 
-import com.assu.server.domain.auth.entity.Member;
+import com.assu.server.domain.member.entity.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.MapsId;
