+++ src/main/java/com/assu/server/domain/auth/service/SignUpServiceImpl.java
@@ -14,8 +14,8 @@ import com.assu.server.domain.auth.repository.SSUAuthRepository;
 import com.assu.server.domain.auth.security.JwtUtil;
 import com.assu.server.domain.common.enums.ActivationStatus;
 import com.assu.server.domain.common.enums.UserRole;
-import com.assu.server.domain.auth.entity.Member;
-import com.assu.server.domain.auth.repository.MemberRepository;
+import com.assu.server.domain.member.entity.Member;
+import com.assu.server.domain.member.repository.MemberRepository;
 import com.assu.server.domain.partner.entity.Partner;
 import com.assu.server.domain.partner.repository.PartnerRepository;
 import com.assu.server.domain.user.entity.Student;
