+++ src/main/java/com/assu/server/domain/member/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.assu.server.domain.member.controller;
+
+public class MemberController {
+}
