+++ src/main/java/com/assu/server/domain/member/converter/MemberConverter.java
@@ -1,0 +1,4 @@
+package com.assu.server.domain.member.converter;
+
+public class MemberConverter {
+}
