+++ src/main/java/com/assu/server/domain/member/dto/MemberRequestDTO.java
@@ -1,0 +1,4 @@
+package com.assu.server.domain.member.dto;
+
+public class MemberRequestDTO {
+}
