+++ src/main/java/com/assu/server/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.assu.server.domain.auth.repository;
+package com.assu.server.domain.member.repository;
 
-import com.assu.server.domain.auth.entity.Member;
+import com.assu.server.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
