+++ src/main/java/com/assu/server/domain/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.assu.server.domain.member.service;
+
+public interface MemberService {
+}
