+++ src/main/java/com/assu/server/domain/member/service/MemberServiceImpl.java
@@ -1,0 +1,4 @@
+package com.assu.server.domain.member.service;
+
+public class MemberServiceImpl {
+}
