+++ src/main/java/com/assu/server/domain/partner/entity/Partner.java
@@ -1,7 +1,7 @@
 package com.assu.server.domain.partner.entity;
 
 
-import com.assu.server.domain.auth.entity.Member;
+import com.assu.server.domain.member.entity.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.MapsId;
